Atualização Criação - Falha - 22/12/2021 - 16:20

+++ backend/src/main/java/com/br/predponta/app/repositories/FalhaRepository.java
@@ -8,6 +8,6 @@ import com.br.predponta.app.entities.Falha;
 
 
 @Repository
-public interface FalhaRepository  extends JpaRepository <Falha, Long>{
+public interface FalhaRepository  extends JpaRepository <Falha, Integer>{
 
 }
